+++ R/lsm_p_ncore.R
@@ -146,14 +146,15 @@ lsm_p_ncore_calc <- function(landscape, directions, consider_boundary, edge_dept
             patch_core <- t(patch_core)
             landscape_labeled <- t(landscape_labeled)
 
+            not_na_patch_core <- !is.na(patch_core)
             # get coordinates of current class
-            points <- data.frame(x = points[which(!is.na(patch_core)), 1],
-                                 y = points[which(!is.na(patch_core)), 2],
-                                 z = points[which(!is.na(patch_core)), 3])
+            points <- data.frame(x = points[which(not_na_patch_core), 1],
+                                 y = points[which(not_na_patch_core), 2],
+                                 z = points[which(not_na_patch_core), 3])
 
-            points$core_id <- patch_core[!is.na(patch_core)]
+            points$core_id <- patch_core[not_na_patch_core]
 
-            points$patch_id <- landscape_labeled[!is.na(patch_core)]
+            points$patch_id <- landscape_labeled[not_na_patch_core]
 
             n_core_area <- table(unique(points[, c(4, 5)])[, 2]) # sth breaking here